my_secure_chat_backend/requirements.txt
django==1.11b1
djangorestframework==3.6.2


django-extensions
ipython
